refactor client view helpers
