+++ view_client.go
@@ -1,17 +1,39 @@
 package emqutiti
 
 import (
+	"strings"
+
 	"github.com/charmbracelet/lipgloss"
 	"github.com/marang/emqutiti/topics"
+	"github.com/marang/emqutiti/ui"
 )
+
+// clientInfoLine renders the connection status and keyboard shortcuts.
+func (m *model) clientInfoLine() string {
+	infoShortcuts := ui.InfoStyle.Render("Switch views: Ctrl+B brokers, Ctrl+T topics, Ctrl+P payloads, Ctrl+R traces.")
+	clientID := ""
+	if m.mqttClient != nil {
+		r := m.mqttClient.Client.OptionsReader()
+		clientID = r.ClientID()
+	}
+	status := strings.TrimSpace(m.connections.Connection + " " + clientID)
+	st := ui.ConnStyle
+	if strings.HasPrefix(m.connections.Connection, "Connected") {
+		st = st.Foreground(ui.ColGreen)
+	} else if strings.HasPrefix(m.connections.Connection, "Connection lost") || strings.HasPrefix(m.connections.Connection, "Failed") {
+		st = st.Foreground(ui.ColWarn)
+	}
+	connLine := st.Render(status)
+	return lipgloss.JoinVertical(lipgloss.Left, infoShortcuts, connLine)
+}
 
 // viewClient renders the main client view.
 func (m *model) viewClient() string {
 	m.ui.elemPos = map[string]int{}
 	infoLine := m.clientInfoLine()
-	topicsBox, topicBox, bounds := m.clientTopicsSection()
+	topicsBox, topicBox, bounds := m.renderTopicsSection()
 	messageBox := m.message.View()
-	messagesBox := m.clientHistorySection()
+	messagesBox := m.renderHistorySection()
 
 	content := lipgloss.JoinVertical(lipgloss.Left, topicsBox, topicBox, messageBox, messagesBox)
